Fix: Parce function call

+++ parcecmd.h
@@ -4,15 +4,15 @@ class TCore;
 class TParceCmd
 {
 private:	
-	int PassName(char* text, uint16_t i);
+	int PassName(char* text, int i);
 	bool isNameChar(unsigned char ch);
 	bool isDigit(char ch);
 public:
 	TCore* Core;
-	int PassNameSpace(char* text, uint16_t i);
-	int PassFuncName(char* text, uint16_t i);
-	bool ParceFunctionCall(char* text, uint16_t i);
-	bool ParceParams(char* text, uint16_t i);
+	int PassNameSpace(char* text, int i);
+	int PassFuncName(char* text, int i);
+	bool ParceFunctionCall(char* text, int i);
+	bool ParceParams(char* text, int i);
 	void Parce();
 
 	TParceCmd(TCore* _core) {
